backend/community/requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
cycler==0.10.0
Django==3.1.1
django-filter==2.3.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-base64==2.0
kiwisolver==1.2.0
kombu==4.6.11
numpy==1.18.2
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
vine==1.3.0
